refactor(e2e-suite): refactor relay server wipe and base fixture file
splits off support methods from suiteBaseFixture.ts
