+++ suite/e2e/package.json
@@ -16,8 +16,7 @@
         "github:create:project": "tsx ./support/reporters/scriptCreateProject.ts",
         "git:bisect": "./scripts/bisect-commits.sh",
         "decode:sol:stake": "tsx ./scripts/decode-sol-staking-account.ts",
-        "docker:suite-sync": "docker compose -f ../../docker/docker-compose.suite-ci-e2e.yml up -d --force-recreate quota-db suite-sync -d && docker compose -f ../../docker/docker-compose.suite-ci-e2e.yml logs -f suite-sync",
-        "docker:suite-sync:stop": "docker compose -f ../../docker/docker-compose.suite-ci-e2e.yml down"
+        "docker:suite-sync": "docker compose -f ../../docker/docker-compose.suite-ci-e2e.yml up quota-db suite-sync"
     },
     "devDependencies": {
         "@currents/playwright": "^1.19.0",
