+++ suite/e2e/support/helpers/evoluClient.ts
@@ -1,7 +1,6 @@
 import { Evolu, SimpleName, getOrThrow } from '@evolu/common';
 import { Upsertable, createEvolu, createOwnerWebSocketTransport } from '@evolu/common/local-first';
 import { expect } from '@playwright/test';
-import { execSync } from 'child_process';
 import { diff } from 'jest-diff';
 import { isEqual, omit, orderBy } from 'lodash';
 
@@ -62,16 +61,6 @@ export class EvoluClient {
         }
 
         return this._evolu;
-    }
-
-    @step()
-    wipeAndRestartServer() {
-        execSync(
-            'docker compose -f docker/docker-compose.suite-ci-e2e.yml up --force-recreate -d quota-db suite-sync',
-            {
-                cwd: '../../',
-            },
-        );
     }
 
     @step()
