+++ suite/e2e/tests/metadata/suite-sync/suite-sync.test.ts
@@ -46,84 +46,72 @@ const expectedOutput = {
     txId: 'aa545d95cf07892e1ae70b40e856b9b476f703e2e20647d0985830fd7b734393',
 };
 
-test.describe(
-    'Suite Sync - Labelling',
-    { tag: ['@webOnly', '@specificFirmware', '@T3W1', '@T3T1'] },
-    () => {
-        test.use({
-            deviceSetup: { passphrase_protection: true },
+test.describe('Suite Sync - Labelling', { tag: ['@webOnly', '@T3W1', '@T3T1'] }, () => {
+    test.use({ wipeEvoluRelay: true });
+
+    test.beforeEach(async ({ onboardingPage, metadataPage }) => {
+        await onboardingPage.completeOnboarding({ keepDebugModeEnabled: true });
+        await metadataPage.setupQuotaManager();
+        await metadataPage.enableSuiteSync();
+    });
+
+    test('Create new labels', async ({ evoluClient, dashboardPage, walletPage, metadataPage }) => {
+        await test.step('Change account label', async () => {
+            await walletPage.openAccount({ symbol: 'btc', type: 'normal', atIndex: 0 });
+            await metadataPage.account.changeLabel({
+                accountId: AccountLabelId.BitcoinDefault1,
+                label: expectedAccount.label,
+            });
+            await expect
+                .soft(walletPage.accountLabel({ symbol: 'btc', type: 'normal', atIndex: 0 }))
+                .toHaveText(expectedAccount.label, { timeout: 30_000 });
         });
 
-        test.beforeEach(async ({ evoluClient, onboardingPage, metadataPage }) => {
-            evoluClient.wipeAndRestartServer();
-            await onboardingPage.completeOnboarding({ keepDebugModeEnabled: true });
-            await metadataPage.setupQuotaManager();
-            await metadataPage.enableSuiteSync();
+        await test.step('Change wallet label', async () => {
+            await dashboardPage.openDeviceSwitcher();
+            await metadataPage.wallet.changeLabel({
+                index: WALLET_INDEX,
+                label: expectedWallet.label,
+            });
+            await expect
+                .soft(metadataPage.wallet.walletLabel(WALLET_INDEX))
+                .toHaveText(expectedWallet.label);
+            await dashboardPage.deviceSwitchingCloseButton.click();
         });
 
-        test('Create new labels', async ({
-            evoluClient,
-            dashboardPage,
-            walletPage,
-            metadataPage,
-        }) => {
-            await test.step('Verify BTC account label is synced', async () => {
-                await walletPage.openAccount({ symbol: 'btc', type: 'normal', atIndex: 0 });
-                await metadataPage.account.changeLabel({
-                    accountId: AccountLabelId.BitcoinDefault1,
-                    label: expectedAccount.label,
-                });
-                await expect
-                    .soft(walletPage.accountLabel({ symbol: 'btc', type: 'normal', atIndex: 0 }))
-                    .toHaveText(expectedAccount.label, { timeout: 30_000 });
+        await test.step('Change address label', async () => {
+            await walletPage.openAccount({ symbol: 'btc', type: 'normal', atIndex: 0 });
+            await walletPage.receiveButton.click();
+            await metadataPage.address.changeLabel({
+                address: expectedAddress.address,
+                label: expectedAddress.label,
             });
-
-            await test.step('Change wallet label', async () => {
-                await dashboardPage.openDeviceSwitcher();
-                await metadataPage.wallet.changeLabel({
-                    index: WALLET_INDEX,
-                    label: expectedWallet.label,
-                });
-                await expect
-                    .soft(metadataPage.wallet.walletLabel(WALLET_INDEX))
-                    .toHaveText(expectedWallet.label);
-                await dashboardPage.deviceSwitchingCloseButton.click();
-            });
-
-            await test.step('Change address label', async () => {
-                await walletPage.openAccount({ symbol: 'btc', type: 'normal', atIndex: 0 });
-                await walletPage.receiveButton.click();
-                await metadataPage.address.changeLabel({
-                    address: expectedAddress.address,
-                    label: expectedAddress.label,
-                });
-                await expect
-                    .soft(metadataPage.address.label(expectedAddress.address))
-                    .toHaveText(expectedAddress.label);
-            });
-
-            await test.step('Change output label', async () => {
-                await walletPage.openAccount({ symbol: 'btc', type: 'normal', atIndex: 0 });
-                await metadataPage.output.changeLabel({
-                    outputId: expectedOutput.txId,
-                    txNumber: Number(expectedOutput.outputIndex),
-                    label: expectedOutput.label,
-                });
-                await expect(
-                    metadataPage.output.outputLabel(
-                        expectedOutput.txId,
-                        Number(expectedOutput.outputIndex),
-                    ),
-                ).toHaveText(expectedOutput.label);
-            });
-
-            await test.step('Verify data are sync to Relay', async () => {
-                await evoluClient.init({ ownerSecret });
-                await evoluClient.expectInTable('account', [expectedAccount], { softExpect: true });
-                await evoluClient.expectInTable('address', [expectedAddress], { softExpect: true });
-                await evoluClient.expectInTable('wallet', [expectedWallet], { softExpect: true });
-                await evoluClient.expectInTable('output', [expectedOutput], { softExpect: true });
-            });
+            await expect
+                .soft(metadataPage.address.label(expectedAddress.address))
+                .toHaveText(expectedAddress.label);
         });
-    },
-);
+
+        await test.step('Change output label', async () => {
+            await walletPage.openAccount({ symbol: 'btc', type: 'normal', atIndex: 0 });
+            await metadataPage.output.changeLabel({
+                outputId: expectedOutput.txId,
+                txNumber: Number(expectedOutput.outputIndex),
+                label: expectedOutput.label,
+            });
+            await expect(
+                metadataPage.output.outputLabel(
+                    expectedOutput.txId,
+                    Number(expectedOutput.outputIndex),
+                ),
+            ).toHaveText(expectedOutput.label);
+        });
+
+        await test.step('Verify data are sync to Relay', async () => {
+            await evoluClient.init({ ownerSecret });
+            await evoluClient.expectInTable('account', [expectedAccount], { softExpect: true });
+            await evoluClient.expectInTable('address', [expectedAddress], { softExpect: true });
+            await evoluClient.expectInTable('wallet', [expectedWallet], { softExpect: true });
+            await evoluClient.expectInTable('output', [expectedOutput], { softExpect: true });
+        });
+    });
+});
